+++ Pod/Classes/Router.swift
@@ -19,6 +19,7 @@ public enum Router {
 	case insert(type: String, fields: [String: Any], authData: AuthData, version: String)
 	case update(type: String, id: String, fields: [String: Any], authData: AuthData, version: String)
 	case delete(type: String, id: String, authData: AuthData, version: String)
+	case describe(type: String, authData: AuthData, version: String)
 	case apexREST(method: HTTPMethod, path: String, parameters: [String: Any]?, headers: [String: String]?, authData: AuthData)
 	case custom(method: HTTPMethod, path: String, parameters: [String: Any]?, headers: [String: String]?, authData: AuthData)
 }
@@ -56,6 +57,9 @@ extension Router: URLRequestConvertible {
 			
 		case let .delete(type, id, authData, version):
 			return try build(authData: authData, method: .delete, path: "/services/data/v\(version)/sobjects/\(type)/\(id)")
+			
+		case let .describe(type, authData, version):
+			return try build(authData: authData, path: "/services/data/v\(version)/sobjects/\(type)/describe")
 			
 		case let .apexREST(method, path, parameters, headers, authData):
 			return try build(authData: authData, method: method, path: "/services/apexrest\(path)", params: parameters, headers: headers)